rename: 텍스트 파싱 로직 파일 위치 이동
