+++ src/entities/text-parsing/index.ts
@@ -1,2 +1,3 @@
 export * from './api'
 export * from './model'
+export * from './lib'
