+++ src/entities/text-parsing/lib/helpers/find-parent-key-by-id.ts
@@ -1,0 +1,15 @@
+import type { AttendanceOptionKey, GetAttendanceOptionsAll } from '@/entities/attendance-option'
+
+export const findParentKeyById = (id: number, attendanceOptions: GetAttendanceOptionsAll) => {
+  let result = null
+
+  Object.keys(attendanceOptions).forEach((key) => {
+    if (
+      attendanceOptions[key as AttendanceOptionKey].detailOptions.some((option) => option.id === id)
+    ) {
+      result = key
+    }
+  })
+
+  return result
+}
