+++ src/entities/text-parsing/lib/index.ts
