--- src/views/text-parsing/lib/helpers/index.ts
@@ -1,2 +1,0 @@
-export * from './transform-attendance-info-to-statistic'
-export * from './parse-text'
